% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule.R
\name{capsule_tree}
\alias{capsule_tree}
\title{Capsule Tree as Nested List / Printable Structure}
\usage{
capsule_tree(store, print = TRUE)
}
\arguments{
\item{store}{A \code{ks_capsule_store}.}

\item{print}{Logical. If \code{TRUE}, print a tree to the console.}
}
\value{
Invisibly, a named list of tree nodes.
}
\description{
Capsule Tree as Nested List / Printable Structure
}
