% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{is_kschat}
\alias{is_kschat}
\title{Test for a \code{kschat} Object}
\usage{
is_kschat(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{kschat}, otherwise \code{FALSE}.
}
\description{
Test for a \code{kschat} Object
}
