% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule.R
\name{ks_capsule_store}
\alias{ks_capsule_store}
\alias{is_ks_capsule_store}
\title{The \code{ks_capsule_store} Class}
\usage{
is_ks_capsule_store(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{ks_capsule_store}, otherwise \code{FALSE}.
}
\description{
A \code{ks_capsule_store} is a named registry of capsules produced by
\code{\link[=as_capsules]{as_capsules()}}, with study-level metadata for persistence and retrieval.
}
