% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsule.R
\name{load_capsules}
\alias{load_capsules}
\title{Load a Capsule Store from a \code{.ksc} File}
\usage{
load_capsules(path)
}
\arguments{
\item{path}{Path to a \code{.ksc} file.}
}
\value{
A \code{ks_capsule_store}.
}
\description{
Load a Capsule Store from a \code{.ksc} File
}
