% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{[[.ks_study}
\alias{[[.ks_study}
\title{Look up an output by id across all types}
\usage{
\method{[[}{ks_study}(x, i)
}
\arguments{
\item{x}{A \code{ks_study}.}

\item{i}{Character scalar output id.}
}
\value{
The matching \code{ks_context}, or \code{NULL}.
}
\description{
Look up an output by id across all types
}
