% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksAI.R
\name{ks_get_option}
\alias{ks_get_option}
\alias{ks_set_option}
\title{Get or Set ksAI Options}
\usage{
ks_get_option(key)

ks_set_option(...)
}
\arguments{
\item{key}{Character scalar. Option name. One of \code{"study_threshold"},
\code{"max_rows"}, \code{"skills_dir"}, \code{"provider"}.}

\item{...}{Named \code{key = value} pairs to set (for \code{ks_set_option()}).}
}
\value{
\code{ks_get_option()} returns the option value. \code{ks_set_option()}
invisibly returns the previous values of the changed options.
}
\description{
\code{ks_get_option()} retrieves a package option; \code{ks_set_option()} updates one
or more options for the current session.
}
\examples{
ks_get_option("study_threshold")
old <- ks_set_option(study_threshold = 30L)
ks_set_option(!!!old)

}
